fuck me

+++ printf/ft_putptr.c
@@ -6,7 +6,7 @@
 /*   By: vvobis <marvin@42.fr>                      +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2024/04/18 17:53:53 by vvobis            #+#    #+#             */
-/*   Updated: 2024/09/17 12:09:29 by vvobis           ###   ########.fr       */
+/*   Updated: 2024/11/10 20:42:58 by marvin           ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -23,5 +23,5 @@ void	ft_putptr(void *ptr, int *count, int fd)
 	}
 	to_print = &ptr;
 	*count += ft_printf("0x");
-	ft_puthex_lower((long)*to_print, count, fd);
+	ft_puthex_lower((unsigned long long)*to_print, count, fd);
 }
